+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
 
+- [apm] fix: Use Performance API for timings when available, including Web Workers (#2492)
+
 ## 5.14.1
 
 - [apm] fix: Check for performance.timing in webworkers (#2491)